+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 * Expose `FunctionAuth`
+* `unquote` no longer raises `IndexError` on empty or single-character strings.
 
 ## 0.28.1 (6th December, 2024)